Added the soft / hard delete option in DeleteAdminModal. Updates backend routes accordingly.

+++ backend/routes/superAdminRoutes.js
@@ -11,7 +11,8 @@ router.get('/admins', async (req, res) => {
   try {
     const admins = await prisma.user.findMany({
       where: {
-        role: { in: [Role.admin, Role.superadmin] }
+        role: { in: [Role.admin, Role.superadmin] },
+        isDeleted: false,
       },
       select: {
         id: true,
@@ -116,10 +117,22 @@ router.patch('/admin', async (req, res) => {
 })
 
 router.delete('/admin', async (req, res) => {
+  const user_id = parseInt(req.query.user_id);
+  const hardDelete = req.query.flag;
+
   try {
-    const user_id = parseInt(req.query.user_id);
-    await prisma.user.delete({where: {id: user_id}});
-    res.status(200).json({message: "Deleted successfully"});
+    if (hardDelete === "true") {
+      await prisma.user.delete({where: {id: user_id}});
+      res.status(200).json({message: "El admin ha sido eliminado por siempre con exito."})
+    } else {
+      await prisma.user.update({where:
+        {id: user_id},
+        data: {
+          isDeleted: true
+        }
+      });
+      res.status(200).json({message: "El admin ha sido eliminado (recuperable) con exito."})
+    }
   } catch (error) {
     console.error(error);
     res.status(500).json({error: 'A server error occurred while deleting the admin'});
